chore: rename reasoning to reasoningText etc
