+++ .changeset/beige-socks-stare.md
@@ -1,0 +1,5 @@
+---
+'ai': major
+---
+
+chore: rename reasoning to reasoningText etc
