+++ examples/ai-core/src/generate-text/amazon-bedrock-reasoning-chatbot.ts
@@ -31,8 +31,8 @@ async function main() {
 
     for (const step of steps) {
       console.log(step);
-      if (step.reasoning) {
-        console.log(`\x1b[36m${step.reasoning}\x1b[0m`);
+      if (step.reasoningText) {
+        console.log(`\x1b[36m${step.reasoningText}\x1b[0m`);
       }
 
       if (step.text) {
