+++ packages/ai/core/generate-text/generate-text-result.ts
@@ -9,7 +9,7 @@ import { LanguageModelRequestMetadata } from '../types/language-model-request-me
 import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
 import { LanguageModelUsage } from '../types/usage';
 import { GeneratedFile } from './generated-file';
-import { ReasoningDetail } from './reasoning-detail';
+import { Reasoning } from './reasoning';
 import { ResponseMessage, StepResult } from './step-result';
 import { ToolCallArray } from './tool-call';
 import { ToolResultArray } from './tool-result';
@@ -28,7 +28,7 @@ The generated text.
   /**
 The full reasoning that the model has generated.
    */
-  readonly reasoning: Array<ReasoningDetail>;
+  readonly reasoning: Array<Reasoning>;
 
   /**
 The reasoning text that the model has generated. Can be undefined if the model
