+++ packages/ai/core/generate-text/stream-text-result.ts
@@ -13,7 +13,7 @@ import { LanguageModelResponseMetadata } from '../types/language-model-response-
 import { LanguageModelUsage } from '../types/usage';
 import { AsyncIterableStream } from '../util/async-iterable-stream';
 import { GeneratedFile } from './generated-file';
-import { ReasoningDetail } from './reasoning-detail';
+import { Reasoning } from './reasoning';
 import { ResponseMessage, StepResult } from './step-result';
 import { ToolCallUnion } from './tool-call';
 import { ToolResultUnion } from './tool-result';
@@ -123,7 +123,7 @@ The full reasoning that the model has generated.
 
 Resolved when the response is finished.
    */
-  readonly reasoning: Promise<Array<ReasoningDetail>>;
+  readonly reasoning: Promise<Array<Reasoning>>;
 
   /**
 The reasoning that has been generated by the last step.
